Rewrite add entry api to use sessions
WIP Delete still is using the old api
