+++ server/dao.js
@@ -1,22 +1,13 @@
 // Define Database Queries
 const db = require('./db')
-const insertEntry = (user, event, lastoccurence, frequency) => {
-  const userId = '(SELECT userid from users where username = $1)'
+
+const insertEntryById = (userid, event, lastoccurence, frequency) => {
   return db
     .query(
-      `INSERT INTO entries VALUES (DEFAULT, ${userId}, $2, $3, $4) RETURNING entryid, userid,lastoccurence, frequency`,
-      [user, event, lastoccurence, frequency + 'days'],
+      'INSERT INTO entries VALUES (DEFAULT, $1, $2, $3, $4) RETURNING entryid, userid,lastoccurence, frequency',
+      [userid, event, lastoccurence, frequency + 'days'],
     )
     .then(res => res.rows[0])
-}
-
-const getEntries = user => {
-  return db
-    .query(
-      'SELECT * FROM entries WHERE userid=(select userid from users where username=$1)',
-      [user],
-    )
-    .then(res => res.rows)
 }
 
 const getEntriesById = user => {
@@ -66,9 +57,8 @@ const getUserIdAndHash = username => {
 }
 
 module.exports = {
-  getEntries,
   getEntriesById,
-  insertEntry,
+  insertEntryById,
   insertUser,
   getOverdueEntries,
   getUserHash,
